Renamed SpeechDecoderViewModel.kt
